requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.9

email-validator>=2.1.0.post1,<3
